Move payouts out of admin controller
